Sign In ready for HTTP integration
I'm now reading the chapter on HTTP calls and considering Volley for an
HTTP Client. I need to Post my new member's information and bother them
about addresses and what not down the line. Truth be told, name, phone
number, email and password should be enough for this program but the
extras will require more information.